Moved excess plotting code from expert data into the plotting code area
